refactor(worker): move cmd/agentdock/local_adapter.go → worker/pool/local.go
LocalAdapter is semantically a local variant of the worker Pool, so it
belongs in the worker module. cmd/agentdock now references it as
pool.NewLocalAdapter.
